-Modificato da double a float i valori di energia in arduino -Modificata la funzione per la comunicazione i2c, ora ci sono due funzioni una per la lettura del bottone singolo e una per la lettura dell'energia

+++ ArduinoNanoI2c/Measure.cpp
@@ -6,8 +6,8 @@
 
 #define REAL_MEASURE		
 
-double 		       EnergyMeasured;
-static double      EnergyInst;
+float 		       EnergyMeasured;
+static float      EnergyInst;
 
 static uint32_t    EnergyTimeCounter;
 #ifndef REAL_MEASURE
@@ -69,7 +69,7 @@ void CalcEnergy() // 200ms c.a.
 
 void EnergyValueSec()
 {
-	EnergyMeasured += (EnergyInst / (double)EnergyTimeCounter);	
+	EnergyMeasured += (EnergyInst / (float)EnergyTimeCounter);	
 	EnergyInst = 0.0;
 	EnergyTimeCounter = 0;
 	EnergyStr = String(EnergyMeasured / 3600.0); // Invio la stringa già formattata per W/h
